Update patch set 3
Patch Set 3:

This one works in the sandbox. I need to push a temp test though to remove a previous file name. Before I was naming the logs compute_1_ovs-vswitchd.log. Then I changed it to ovs-vswitchd_compute_1.log so that the ovs-vswitchd log files were grouped together. But these vms workspace kept the old files around and show up in the logs. So need to push a job to delete them.

Patch-set: 3
Label: Verified=0