Update patch set 1
Patch Set 1: Code-Review+2

This IS hiding a bug from pure passrate point of view, but on the other hand it allows us to distinguish performance bugs from worse ones. The deciding factor is that I think combination of 3GB heap and 4GB RAM is currently not stable enough.

Patch-set: 1
Label: Code-Review=+2