Update patch set 1
Patch Set 1: Code-Review-1

something worked.  look here:

https://logs.opendaylight.org/sandbox/jenkins091/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-carbon/3/archives/

the idea is that I'm trying to create a folder in archives/ for control, compute_1 and compute_2 and then put all the relevant logs there.  It seems to work for compute_1 and compute_2 and the logs show up under those folders.  I think I'm doing the same thing for
compute but the logs don't end up in a "compute" folder.  They end 
up with their root in archives/  

help?

JamO

Patch-set: 1
Label: Code-Review=-1