adding odl-netconf-topology install feature
the purpose of this change will be a step by step removal of
odl-netconf-connector-all(ssh) feature from the jobs
step 1) adding odl-netconf-topology (present change)
step 2) replace connector functionality in integration/test suites
step 3) removal of odl-netconf-connector-.. from installed features

installing of odl-netconf-topology should not affect any jobs. No
collisions are expected in B and C. Be jobs may log some exceptions
as the feature is not existing, but this should be harmless to the
functionality.

Change-Id: Ibd39888516fcd240e864a6c9befc6ae57a398ac4
Signed-off-by: Peter Gubka <pgubka@cisco.com>