Update patch set 4
Patch Set 4:

Luis pointed out our mistake.

The archives dir is not there yet when we are doing this log collection.
It is created in the "integration-csit-archive-build" step which runs
after this.

I want to get Thanh's opinion on us doing it this way, vs letting
the archive build macro do it.  To do it that way, we need to collect
the logs like we currently are and keep them in ${WORKSPACE} and then 
modify the archive build macro to include these files in this line:

https://github.com/opendaylight/releng-builder/blob/master/jjb/integration/integration-macros.yaml#L512


I'll updated this patch to first create the archives dir (instead
of dummy) which we know will work.

Ok with you Thanh, Luis, Vratko?

Patch-set: 4
Label: Code-Review=0