Update patch set 6
Patch Set 6:

I've requested the newton version of this be created as a VM in the sandbox.  When that happens we can try that image on both the newton
and mitaka jobs.  If that works, then we can just go with a single
image.  That would mean a small refactor here removing the mitaka json and renaming the newton json to just devstack-pre-pip

at least that's how I see it for now.  waiting on the releng folks to provide the image at this point.

Patch-set: 6