Update patch set 7
Patch Set 7:

the newton image is in the sandbox now.  I've tested it on the newton 
jobs and it looks to be aprox 10m faster.  here is with [0] and 
without [1].  But, the other question is if this image would work on 
our mitaka job.  Mitaka was broken for the past 24h or so, but I think
it should be fixed now.  I've kicked off a mitaka job [2] with this
newton image and we can see.

If the mitaka job is ok with the newton image, don't we want to remove
the entire devstack-pre-pip-mitaka.json and just have one?  And if so,
do we even care to label it as newton at that point, or just
devstack-pre-pip is good enough?

[0] https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-jamo-upstream-transparent-carbon/3/
[1] https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-jamo-upstream-transparent-carbon/2/
[2] https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-boron/1/

Patch-set: 7