Update patch set 1
Patch Set 1: Code-Review+1

Alon, it isn't a lot since it just adds argument tracing to the keyword calls. That is the reason I wanted it - so that we could see the args since they are so difficult to follow sometimes - is this a list, is it empty...

I was debating adding it by default to see the args for a while. I think we can add that as a follow on patch so we can easily revert when we want to change it back. You simply set the robot-options field to be '-L TRACE' in the yamls we want.

A quick look at the log size, though, did show the log doubled. A typical test without logs had a 8mb log.html and with TRACE level it became 14Mb. TRACE is the highest level and captures more than the args but I didn't see much else in the log.html. There are other levels like DEBUG that capture less.

Patch-set: 1
Label: Code-Review=+1