Update patch set 3
Patch Set 3:

> > (1 comment)
 > >
 > > I am unsure about the RECLONE logic, especially since the comment
 > > says it needs to be removed.  I had assumed it was there just for
 > > debugging
 > > purposes in the sandbox.
 > >
 > > Merging this now, as I think it gets our mitaka jobs stacking
 > > again, and the current state is broken.
 > 
 > 
 > Jamo,
 > 
 > The RECLONE=no was added to avoid cloning the "requirements" and
 > the "python-openstacksdk". The changes need to be reverted if the
 > fix is made in the individual projects.

Venkat, when you have time could you update the comments to be more clear so people will know better what's going on if/when we need to revert/refactor?  Like when/how do we know if the fixes are made in the 
individual projects (if that's even possible), etc.

Patch-set: 3