Update patch set 1
Patch Set 1:

Figured out a way how we can test if it is working or not with Genius CSIT. Since it takes only 30 mins to run and is not facing any issues, makes a better candidate for this. Probably add triggers for both?

Patch-set: 1
Reviewer: Gerrit User 3276 <3276@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0