add core feature to ofp patch test
There is an implicit "all" already built in [0] to our patch test
that will run *all* of the CSIT jobs in the .yaml and does not
need an "all" feature defined. I'm not sure how that {csit-list}
gets created, but it's how it's working in practice, and seems
handy.

So, if this is merged the new gerrit keyword would be test-openflowplugin-core
BUT the old test-openflowplugin-all would still give the same results for
now. However, this will let others (like in this patch [1]) be able to add
more interesting scenarios (e.g. test-openflowplugin-netvirt) and if that
comes along doing test-openflowplugin-all will now include those new ones.

[0] https://github.com/opendaylight/releng-builder/blob/master/jjb/integration/integration-templates.yaml#L1248
[1] https://git.opendaylight.org/gerrit/#/c/50033/

Change-Id: I537a835f0c88998ad73c32f624a32c2017baf90c
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>