Update patch set 1
Patch Set 1:

(4 comments)

We had some improvements planned, which were not implemented because they were not backward compatible. This is a backward-incompatible change, so it is an opportunity to squeeze the improvements in.

But maybe it would be better to separate that improvements from this change, so my vote stays at zero.

Patch-set: 1