Update patch set 1
Patch Set 1:

Granted it is true that we do not produce a perfect output of what was built in our final code due to the fact that we have an autorelease build. In a perfect world projects should be releasing separately and we consume them via a project that pulls the release bits from each project from Nexus.

Patch-set: 1