let number of hypervisors be 1 in the case that it's a single node openstack deployment
a recent patch [0] was merged to double check that devstack has brought up the correct
number of hypervisors in the multinode openstack deployments. this patch didn't take
in to account that some jobs (e.g., vtn's openstack job) is doing control and compute
in a single node. this should fix that.

Change-Id: Ib7fd8a8c1d854199c0b501d2f3f71104c1914d87
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>