Update patch set 2
Patch Set 2:

Yes this is fine. "8 GB General Purpose v1" in rackspace public cloud is a 8 core 8 gb system. The cores part of the image only matters in the private cloud but none of the csit jobs currently run in the private cloud so we're good. We can tweak the flavors if we ever decide to run csit in private cloud again.

Patch-set: 2
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0