Update patch set 14
Patch Set 14:

The pybot line is not revoked yet. Though the changes may be simple. It could affect a lot of jobs as this script is used by multiple jobs (verify jobs included). Are we intending to make this file as only script to deploy ODL (1node and 3node)? If so, we need to do more testing of all the 1node and 3node jobs using this script.

Patch-set: 14
Label: Verified=0