Update patch set 1
Patch Set 1:

Do we need a new node for the hwvtep or can we just add hwvtep to an existing node and reuse the current netvirt csit? Seems like hwvtep emulator just needs to be on a node with ovs. The current tests spin up a odl node, openstack control and two compute nodes. We could add the hwvtep to the odl node.

The hwvtep code is part of netvirt so seems like it should be tested as part of it's suite.

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0