Update patch set 14
Patch Set 14: Code-Review-1

> > The pybot line is not revoked yet. Though the changes may be
 > > simple. It could affect a lot of jobs as this script is used by
 > > multiple jobs (verify jobs included). Are we intending to make
 > this
 > > file as only script to deploy ODL (1node and 3node)? If so, we
 > need
 > > to do more testing of all the 1node and 3node jobs using this
 > > script.
 > 
 > Hi,
 > I did check the verify job for 1 node in order to check the pybot
 > line:
 > 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-legacy-csit-verify-1node-openstack-basic/1/
 > It seems ok to me, what do you think?
 > 
 > For your second question, I think that we would  like to make it
 > the only file for both 1/3 nodes, but I think we should do it step
 > by step (that's why I didn't remove all 3 nodes bash file).
 > For this step I tested all jobs that using this template and run 1
 > verify job to see the pybot line.

Sandra, I think the easiest thing here is just to fully revert the pybot line
in the include-raw-integration-deploy-controller-run-test.sh suite so that
it has no changes in this patch. If you look up around line 232 you can 
see that when this script is run with non-openstack jobs, it will exit 
there and that pybot command is never run. it is run in all of our non-openstack
jobs though. I think we don't care about those jobs for this patch right?
saves time/testing to just revert it.

Patch-set: 14
Label: Code-Review=-1