Update patch set 1
Patch Set 1:

> Do we need a new node for the hwvtep or can we just add hwvtep to
 > an existing node and reuse the current netvirt csit? Seems like
 > hwvtep emulator just needs to be on a node with ovs. The current
 > tests spin up a odl node, openstack control and two compute nodes.
 > We could add the hwvtep to the odl node.
 > 
 > The hwvtep code is part of netvirt so seems like it should be
 > tested as part of it's suite.

if this is possible, it would be better as it will help save 
resources. 

what is going to be done to the node with this hwvtep emulator?
does it have to be a node that is not doing the normal ovs stuff
with openstack? meaning, we can't do hwvtep tests with it
running on the compute or control openstack nodes?

either way, doing it on the odl node seems like an option as well.

btw, first glance at this patch didn't seem like anything major
was wrong.

Patch-set: 1
Label: Code-Review=0