Update patch set 1
Patch Set 1: Code-Review-1

We also need to update the image in Jenkins under the OpenStack configuration. The changes here only apply to CSIT jobs.

Also we should search for all the projects overriding the defaults as there are several projects doing that and ensuring the relevant images get updated everywhere.

Patch-set: 1
Label: Code-Review=-1