Enable JJB weekly job
Long ago we disabled this job because we were having issues with the job
always timing out and failing. JJB has since introduced things like
parallel updates which should make this job more stable. It is good to
run something like this once a week to reset our jobs and ensure
there's no discrepency with what is in releng/builder vs the jobs in
production.

Change-Id: I11ca998c907a219c0be05d2473737a37ad6362d9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>