adding change discovery tool to run and display merged patches from the running distribution
this is hopefully a temporary place to run this tool. Ideally I would like to expose it at
the robot level and/or create jenkins artifact files with the relevant info collected
by changes.py. This will take more work, and this is an easy way to start taking advantage
of this new tool.

Change-Id: If3cd5c74b4591be16a0f0518402a4c92ad4c5fec
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>