Update patch set 1
Patch Set 1:

as the commit message says, I want this to be a temporary addition.
But it's quick and easy and will immediately put this new tool
to good use and at least give us the info in the console log, like
you can see in this sandbox run:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-carbon/47/consoleFull

search for the string "Using these options" to see it.

I'll work on running this from robot next and providing an easy
way for all projects to use it.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0