Add job to generate an image list for docs
With Heat now available we need a way to automate the list of available
images projects can use in their jobs in the docs. This creates a job
that will pull from openstack a list of available public images and
pushes a patch automatically to Gerrit to update the relevant
documentation page. The script will automatically detect if an unmerged
image update patch already exists and update existing patch if available
otherwise create a new one.

Change-Id: If5ea2c53b21b06ab1ed5d6c9c17f30356a930dc4
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>