Update patch set 1
Patch Set 1:

Turns out Jenkins installs Maven using wget in a custom shell script. So editing .wgetrc appears to be the only way to affect that part of the build.

Haven't figured out how to inject into unzip a way to be less verbose though but this seems to make things a little more quiet at least.

Patch-set: 1