Update patch set 1
Patch Set 1: Code-Review-1

This is good. We use spinup scripts temporarily until we can rebuild images with the installs built in so those packages should have been moved over.

We have to be careful here though as we cannot merge this patch until images are rebuilt first. Perhaps we should split the patch into 2 patches, one to update packer and a 2nd one that depends on this one to remove from the spinup scripts parts.

We have to make sure that the images are updated and replaced successfully before we can remove the spinup script calls.

-1 to so that people are aware that if we merge this too hastily we could break builds.

Patch-set: 1
Label: Code-Review=-1