Update patch set 2
Patch Set 2:

One option is to have the merged jjb/ directory assume it is used on Sandbox and let jobs related to configuring real RelEng jobs apply a patch to change a macro name just before running jenkins-jobs.

It would be ugly, non-intuitive (unless well-documented) and hard to debug, but it would work.

Patch-set: 2