Update patch set 8
Patch Set 8:

This is cool, since we are having a job for updating the images. would be good also delete images which are not used or public and possibly keep only the last two used images, in-case we need to role back, if not create a separate job for that.

Patch-set: 8
Reviewer: Gerrit User 6258 <6258@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0