Update patch set 2
Patch Set 2:

Hi Guy,

2 weeks ago we upgraded to Config File Provider plugin 2.15.5 and OpenStack Plugin 2.17 (but had to revert back to 2.11). If you have not yet upgraded to these versions I would highly recommend you DO NOT upgrade. There's an issue with Config File Provider and OpenStack plugins that cause the "Manage Files" page to render a blank page making it so we can no longer edit managed files. This is what's blocking us from moving forward at the moment.

While upgrading OpenStack plugin to 2.17 with a workaround would allow us to edit the files again there's a different bug in 2.17 that prevents us from booting VMs correctly  which is why we had to revert back to 2.11. Until this is fixed in the OpenStack plugin we're actually blocked on the ability to edit our managed files.

Patch-set: 2