Update patch set 2
Patch Set 2:

Hey Guy, just wanted you to be aware of this since you were the original contributor of this file. If you need this we'll need to find a different way to do it as the current way does not work.

Patch-set: 2
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0