Update patch set 5
Patch Set 5: Code-Review+1

ok, bug fixed.

here is a sandbox using this patch:
https://jenkins.opendaylight.org/sandbox/job/distribution-merge-carbon/17/console

two things to note:

1) the job is a red dot because the sandbox can't upload artifacts to
   nexus so it fails there. This new functionality uses || true so
   even if there is trouble the merge job would not fail because of
   it.

2) the sandbox job is using an unmerged version of the distcompare
   tool. until that patch is merged, we wont get any benefit from
   this patch to use it. but, because of 1) I don't think we have to
   wait to merge this.

OH, the log server now will have this:
https://logs.opendaylight.org/sandbox/jenkins091/distribution-merge-carbon/17/archives/dist_diff.txt.gz

scroll to the bottom to see the info that would be important. there
is a TODO in this patch and it would be a good addition to make things
easier to read/see, but that can be later.

Patch-set: 5
Label: Code-Review=+1