Update patch set 2
Patch Set 2:

> > (1 comment)
 > Since we are doing the auto-release builds everyday!
 > I think we could use the --after=yesterday flag,
 > 
 > ex: `git log --pretty=format:"%h %s" --graph --after="yesterday"`
 > which will give the summary of all commits since yesterday.

Not sure how helpful that is because if we run the job more than once we should have a different change list as we want to know what changed between builds. Jenkins actually gives us a better version of this via the Changes page as it tells you what changed since last build.

Patch-set: 2