Update patch set 7
Patch Set 7:

> (1 comment)

Hi Jamo, Thanh,

This has been taken care in the include-raw-integration-deploy-openstack-run-test.sh "line 152" that if this variable is not available (or not defined) it would not change the existing behavior.
Should that be ok?

I have also attached a passlog with the integration master (for a job created with another openstack yaml) and is full pass.

Re-attaching the log here.
https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-upstream-transparent-carbon/2/


I shall re-bases and fix the merge issues and the extra space in commit.
Awaiting your response.

Thanks,
Abbas

Patch-set: 7