virtualenv workaround to avoid 128 char length limitation
This fix is based off a similar solution that OpenStack used in [1].
The solution is to call the python scripts via python
itself. Doing this ignores the shebang line in the scripts.

This is only a partial fix and covers what is in opendaylight-infra-stack.sh.
Any other place that has the virtualenv also needs the fix.
Another example is in opendaylight-infra-copy-ssh-keys.sh
and the openstack python script is called.

[1] https://review.openstack.org/399754

Change-Id: Id38275dd9352bd5cdbcddf7a096b04902ee496c6
Signed-off-by: Sam Hague <shague@redhat.com>

virtualenv workaround to avoid 128 char length limitation

Change-Id: I3b9cbf1c1a8dc49842fb615c166c4a383aebd7ec
Signed-off-by: Sam Hague <shague@redhat.com>

virtualenv workaround to avoid 128 char length limitation

Change-Id: I3aaad756106adef07e50454ce344d803d7c92d11
Signed-off-by: Sam Hague <shague@redhat.com>