Update patch set 2
Patch Set 2:

Yeah, PYTHON and OPENSTACK should move to a lib and then used by the scripts. My main goal with this here was to show a possible solution. The hwvtep job hit the limit and we just shortened the job name, but the other jobs are sitting around 110 chars so we will likely hit it again. Could also shorten -openstack to just -os. in the job names there is also an openstack piece we can shorten to os. csit to it...

I saw other solutions like the below which monkeys with the shebang limit. That seemed just as painful.

#!/bin/bash
script="$1" 
shebang=$(head -1 "$script")
interp=( ${shebang#\#!} )        # use an array in case a argument is there too
# now run it
exec "${interp[@]}" "$script"

Patch-set: 2