Allow topic branches to use distribution-check
The current distribution-check job assumes that the gerrit branch is known
to the integration/distribution branch, i.e. a patch on stable/boron
would build a distribution for stable/boron. This fails when using a topic branch
where integration/distribution only has the upstream branches, stable/boron and master.

This patch adds a variable: distribution_branch and sets the default to GERRIT_BRANCH.
For topic branches, they can set distribution_branch in their project.yamls to be
the right branch.

Change-Id: Icb39a2de15d66f3998daa68d3a40b95362029ae6
Signed-off-by: Sam Hague <shague@redhat.com>