Update patch set 3
Patch Set 3: Code-Review+1

well the happy path with this change is ok, so not breaking anything there.

https://jenkins.opendaylight.org/sandbox/job/aaa-csit-1node-authn-all-boron/1/console


If there is some trouble with this change, I think it would show
up in the path where we are trying to recreate the stack after
deleting the failed stack. But, if that path is hit, then the job
was going to be a red dot anyway.

so, I think the risk of merging is low, but we do need to search 
and find the examples of CREATE_FAILED so we can see if this is
helping us or not.

Patch-set: 3
Label: Code-Review=+1