Update patch set 8
Patch Set 8: Code-Review+1

I think this is ready.

using this script in the sandbox, here is a normal stack start:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-boron/602/console


here are two examples when the first stack create failed, and the
retry worked:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-boron/599/console

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-boron/583/console


note: while working on this, I noticed (maybe only twice in the 600+
jobs) that a delete was stuck in DELETE_IN_PROGRESS. Eventually that
should timeout and go to DELETE_FAILED. In that case, there is logic
to try a "stack abandon" as a last ditch effort.

Patch-set: 8
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1