Update patch set 8
Patch Set 8: Code-Review+2

I don't like that we're printing stack list so many times as it spams the console but we can fix that in a later patch if it gets too spammy. Let's merge this and see how much further it gets us to a successful integration-distribution-test job run.

Patch-set: 8
Label: Code-Review=+2