Update patch set 1
Patch Set 1:

Not sure if this actually works but I rebuilt a CentOS machine with this and it seemed to fix things. We can monitor and see if we get kernel panic issues in future image builds.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0