Update patch set 13
Patch Set 13:

> fine to merge, but want to double check on the failure in the
 > new test suite. This is a bug in ODL and/or networking-l2gw right?
 > 
 > if so, will it be fixed soon?

Hi Jamo,

The failure in the new suite is purely because of the order of testcases I am running. But I have verified that the configuration testcases are all working fine. So there are no bugs in ODL as such.

I shall push the test suite in a different gerrit once my testing is completed.

Have I answered your query? If so can you please help to merege this?

Thanks,
Abbas

Patch-set: 13