Update patch set 2
Patch Set 2: Code-Review-1

Hold on with this before merging - see first comment.
For the tests to actually work, we need an integration/test patch.
Also it seems like the appdirs issue is failing stacking (we hit this in newton and mitaka but Venkat did something to workaround - seems like we're doing something wrong)

Patch-set: 2
Label: Code-Review=-1