Update patch set 1
Patch Set 1:

this was created in order to deal with this patch:

https://git.opendaylight.org/gerrit/#/c/52525/3

although the 52525 patch may not been needed in the end, as it 
exists now, this 52562 patch should still be ok. it allows
for the command line scripts to take arguments. without this
change every whitespace separated element on each real line is
looped over in turn.

Patch-set: 1