Make link to tar from BUNDLEURL, use for build-rpm
RPM specfiles typically use tarball build inputs, not zips. When
triggering from autorelease, we need to convert the BUNDLEURL link to
the zip-archive distro into one for the tarball distro.

Could also do this by creating a new variable parallel to BUNDLEURL for
passing a link to the tarball, something like BUNDLETARURL. I started on
this in [1] but think it's not worth the complexity, unlikely to be
reused.

[1]: https://git.opendaylight.org/gerrit/#/c/52693/

This fixes:

trello.com/c/qcwHU9u1/446-fix-build-rpm-jobs-triggered-by-autorelease

Change-Id: Ic9b4a998a4f823f4a3a3b1ef3ac0f604cf03565f
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>