Update patch set 3
Patch Set 3: Code-Review+2

> One of my motivations is that "master" is not really a stream.

I've always thought of STREAM as a generic term to mean anything. In ODL it means elements of the periodic table but in other projects it could mean a version number or just line up with their branches.

Anyway it's all matter of opinion, I guess let's just get this merged since there's no harm in it.

Patch-set: 3
Label: Code-Review=+2