Add workaround for mkinitrd failing on CentOS7
Sometimes yum update fails to produce a working grub config as the
kernel install fails to build mkinitrd. This workaround introduces a
loop to reinstall the kernel if grub.conf is missing the initramfs line
which is an indication of the mkinitrd issue. The new function will
cause a packer failure via `exit 1` if the workaround did not work.

Change-Id: Ib498a610fdd6f082575495d1b56a57b5f1802321
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>