Simplify patches tarball creation
We should be pulling in everything in the PATCHES directory (especially
taglist.log) and it does not hurt to also pull in the diff files as a
backup for the bundle files.

Also fixes issue where the tarball contains the path w/jenkins-host/...
we just need it to include a patches directory.

Change-Id: I6d59eaced8ee34dce04fdf74d244196d227ef4cb
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>