Update patch set 1
Patch Set 1: Code-Review+2

We should probably figure out a better way to handle this at the job template level because this makes the archive script archive for every job regardless of if they need it or not. But in the meantime since we're already doing it this way let's get it merged.

Patch-set: 1
Label: Code-Review=+2