Add job to automatically cleanup old images
Depends on lftools>=0.0.8. Adds a jjb job to automatically cleanup
non-public images in the Private Cloud and Rackspace Public clouds.

Criteria:

    - Uses the sandbox account and checks only the sandbox tenant
    - Images 30 days or older
    - Non-public images

In ODL we mark images public when we want to allow the community to use
them. These are officially supported images that has passed testing
typically in the sandbox.

Change-Id: Id4ce204393408a087739ea18d1dd9700d346ad30
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>