Update patch set 1
Patch Set 1:

> Sorry Jamo, but I don't fully understand. This patch has to do with
 > "sfc" project and the patch you mention is related to "aaa" project

yeah, it's not technically related, but somehow this patch is a
child of 52942 (see the "changes" box in the top right of this 
gerrit UI?). because of that, I can't merge (like the submit button
is greyed out) this patch. I looked at 52942 and that one is a new
job for AAA, and that new job will waste jenkins resources if it's
merged and running before the system test code is merged, which is
why I mentioned 52735.

sorry for not being too clear.

Patch-set: 1