Add option to not skip git.properties
The q(uick) profile skips the git-commit-id-plugin.
This patch adds the skip override so that the plugin
wil run and the git.properties file included in the jars
so that the distchanges tool works properly.

There is also some cleanup to make the distribution builders
be more consistent and use the quick profile.

Change-Id: I5476f8c911ed7899a78577a6e61e1e88b620b747
Signed-off-by: Sam Hague <shague@redhat.com>