Trigger verify on commit-message-only changes
Some project check commit message formatting in -tox- jobs.

This change is global across all Gerrit-triggered jobs,
so all other verify-type jobs are triggered as well.
That is to avoid false Verify+1 votes
when the previous patch set failed on a code-sensitive job.

Also, values without variables do not need to be quoted.

Change-Id: I8eb8cb21c99acd9d2f92a76c941312ce21efa89c
Signed-off-by: Vratko Polak <vrpolak@cisco.com>