Update patch set 6
Patch Set 6: Code-Review+1

I think its good to have this committed keeping record of how to do it, we've been asked about this a lot. No need for a perioidic job to actually run with it in the background.

Running again in sandbox to verify it works ok.
https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-alonko-carbon/9

Jamo thoughts?

Patch-set: 6
Label: Code-Review=+1