Update patch set 5
Patch Set 5:

> giving it another try - https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-upstream-stateful-alonko-carbon/5

nothing obvious stood out to me in the 45m I spent digging through the openstack logs. the
VM is just stuck in "scheduling", even after maybe 90s from being requested. I did see 
a bunch of hits when I asked google for "ocata instance stuck scheduling" but nothing 
pointed me to anything that was obvious.

trying again:
https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-upstream-stateful-alonko-carbon/6/

this time with a 5m polling on the VMs to be active.

Patch-set: 5