Update patch set 8
Patch Set 8:

I am a little lost here but my suggest for creating separated template was in case  you cannot come up with what you want with existing. For the sanity name, Jamo is right the test has to be 100% stable, one way to achieve this is to temporary exclude known bugs and unstable test, another option is to wait until the test gets stable (hopefully before we release carbon) and then change the name to sanity.

Patch-set: 8
Label: Verified=0