Update patch set 7
Patch Set 7:

I think I am OK with using a parent job. It is nice to get rid of the triggers. Would this also structure it so that we don't really end up with a job called high-freq? Seems goofy to have the exact same running tests in two different jobs just because one runs more often.

Patch-set: 7