Handle bundle-vars directly in dist-compare script
instead of using the get-bundle-vars script as a build step
which can fail when there is really no "last" build (like
when cutting a new branch) we can figure out those variables
directly in the compare-distribution script which is using
set +e. If there is any trouble with the compare-distribution
script, we don't abort the build.

Change-Id: I8442149ae4c0f3ddb3f69b73dbac5df3c948a4b8
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>