Update patch set 2
Patch Set 2:

thoughts on this approach?

I used the name "high-frequency" to be generic, but don't care
if we call it whatever seems better, if anything.

also, this is not changing our actual csit job name, so that will
still be picked up by the nightly triggers we already have. This is
a good thing, imho.

Patch-set: 2
Label: Verified=0