Make patch-test runnable without passing REFSPEC
Gerrit will always pass GERRIT_REFSPEC and override this property if a
job is triggered by Gerrit Trigger. Sometimes it's nice to be able to
just click build on a job to test it so configure a default refspec
which allows this. It is still possible for the user to override and set
a specific GERRIT_REFSPEC too when they manually start the job.

Change-Id: I1b200b97d63444bcdabfb2f9637dc8d8171337c9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>