Update patch set 1
Patch Set 1:

> (4 comments)

Using the approach to pass command options to cmds is not recommended because 1. $SUBJECT would have spaces and -s without the quotes around var substitution would not include the fully string in the subject line and just the first word. I think its more simpler to use a if else clause in this case. 

http://mywiki.wooledge.org/BashFAQ/050

Patch-set: 1