Add .npmrc
Make any npm commands infra provided .npmrc. Currently this file
contains a registry configuration to make npm use Nexus as a npmjs
registry which is setup as a proxy of upstream npmjs' repository. This
allows us to be more resilient to upstream infra issues that are out of
our control.

Change-Id: I4c8451aa66921fecffb9f968d7c7bb42214dde2a
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>