Fix distribution-verify job back to freestyle
The big patch https://git.opendaylight.org/gerrit/53797 accidently
converted the distribution-verify job back to a Maven project type and
causes builder-jjb-merge to fail because it is not able to update a job
that's of a different type.

Since we don't want to be reverting jobs back to Maven project types
anyway this patch updates the job to use freestyle again.

Change-Id: I95f50ad683b6f2486b246b47e87f5f1f1305f005
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>