Update patch set 2
Patch Set 2:

I do not quite agree Vratko, I think not disturbing the patch verification at this moment is same or more important than motivating projects to add their feature repos to Karaf 4 distribution. This temporary patch is precisely to avoid any impact on projects patch verifications routines. If this routine is to be changed we should: 1) inform about the change (e.g. not possible to verify on nitrogen until karaf 4 is ready); and 2) explain how to override the behavior (e.g. changing the project jjb files).

Patch-set: 2