Update patch set 1
Patch Set 1:

Jamo, the issue in your case is that you are using a distribution created in jenkins sandbox and current script is not ready for that. Your patch may fix something but may alco break the distribution coming from autorelease so we have to be careful here. Let me check at Vratko's run too and come with something safe.

Patch-set: 1